Reduce the duplication of `RegExpShim` and `getFileScanRDD`[databricks] (#7123)
* Clean the duplication of RegExpShim and method "getFileScanRDD"

Move the single method into SparkShims and remove the RegExpShim files.
