--- sql-plugin/src/main/312-nondb/scala/com/nvidia/spark/rapids/shims/RegExpShim.scala
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2022, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nvidia.spark.rapids.shims
-
-object RegExpShim {
-  // Handle regexp_replace inconsistency from https://issues.apache.org/jira/browse/SPARK-39107
-  def reproduceEmptyStringBug(): Boolean = true
-}
