+++ sql-plugin/src/main/312-nondb/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -27,6 +27,8 @@ object SparkShimImpl extends Spark31XShims {
 
   override def hasCastFloatTimestampUpcast: Boolean = true
 
+  override def reproduceEmptyStringBug: Boolean = true
+
   override def getParquetFilters(
       schema: MessageType,
       pushDownDate: Boolean,
