+++ sql-plugin/src/main/313/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -45,4 +45,6 @@ object SparkShimImpl extends Spark31XShims {
   override def hasCastFloatTimestampUpcast: Boolean = true
 
   override def isCastingStringToNegDecimalScaleSupported: Boolean = true
+
+  override def reproduceEmptyStringBug: Boolean = true
 }
