+++ sql-plugin/src/main/320+/scala/com/nvidia/spark/rapids/shims/Spark320PlusShims.scala
@@ -471,4 +471,6 @@ trait Spark320PlusShims extends SparkShims with RebaseShims with Logging {
     override def convertToGpu(): GpuExec =
       GpuBatchScanExec(p.output, childScans.head.convertToGpu(), runtimeFilters)
   }
+
+  override def reproduceEmptyStringBug: Boolean = false
 }
