+++ sql-plugin/src/main/320/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -19,28 +19,14 @@ package com.nvidia.spark.rapids.shims
 import com.nvidia.spark.rapids._
 import org.apache.parquet.schema.MessageType
 
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.AttributeReference
-import org.apache.spark.sql.execution.datasources.{DataSourceUtils, FilePartition, FileScanRDD, PartitionedFile}
+import org.apache.spark.sql.execution.datasources.DataSourceUtils
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFilters
-import org.apache.spark.sql.types.StructType
 
 object SparkShimImpl extends Spark320PlusShims
     with Spark320PlusNonDBShims
     with Spark31Xuntil33XShims
     with Spark320until340Shims {
   override def getSparkShimVersion: ShimVersion = ShimLoader.getShimVersion
-
-  override def getFileScanRDD(
-      sparkSession: SparkSession,
-      readFunction: PartitionedFile => Iterator[InternalRow],
-      filePartitions: Seq[FilePartition],
-      readDataSchema: StructType,
-      metadataColumns: Seq[AttributeReference]): RDD[InternalRow] = {
-    new FileScanRDD(sparkSession, readFunction, filePartitions)
-  }
 
   override def getParquetFilters(
       schema: MessageType,
@@ -59,4 +45,6 @@ object SparkShimImpl extends Spark320PlusShims
   }
 
   override def isCastingStringToNegDecimalScaleSupported: Boolean = false
+
+  override def reproduceEmptyStringBug: Boolean = true
 }
